+++ Sources/CyPlayer.h
@@ -160,9 +160,9 @@ public:
 	int calculateTotalCityHealthiness();
 	int calculateTotalCityUnhealthiness();
 
-	long long getFinalUnitUpkeep();
+	int64_t getFinalUnitUpkeep();
 	int calculateUnitSupply();
-	int calculatePreInflatedCosts();
+	int64_t calculatePreInflatedCosts();
 	int calculateInflationRate();
 	int calculateInflatedCosts();
